requirements.txt
langchain==0.2.13
langchain-core==0.2.35
langchain-community==0.2.12
langchain-openai==0.1.23
python-dotenv==1.0.1
faiss-cpu==1.8.0
tiktoken==0.7.0
